rename route name, model name
